requirements.txt
alembic==1.8.0
fastapi==0.78.0
SQLAlchemy==1.4.49
uvicorn==0.17.6
dlib==19.24.2
filelock==3.12.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mpmath==1.3.0
networkx==3.1
numpy==1.25.1
opencv-contrib-python==4.8.0.74
opencv-python==4.8.0.74
Pillow==10.0.0
sympy==1.12
torch==2.0.1
typing_extensions==4.7.1
mysql-connector-python==8.0.17
